Safari-609.1.11.1/LayoutTests/media/W3C/video/src/src_reflects_no_value.html

<!doctype html>
<html>
 <head>
  <title>video.src - empty if no source</title>
  <script src="../../w3cwrapper.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-src">spec reference</a></p>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
  document.getElementById("v").src,
  "",
  "videoElement.src should be empty if there is no source");
});
  </script>
 </body>
</html>
