LayoutTests/media/W3C/video/src/src_reflects_no_value-expected.txt

video.src - empty if no source

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" is ""

TEST COMPLETE
spec reference
